+++ belita-cloud-service-order/pom.xml
@@ -40,6 +40,12 @@
             <artifactId>spring-cloud-starter-alibaba-nacos-config</artifactId>
         </dependency>
 
+        <!-- sentinel -->
+        <dependency>
+            <groupId>com.alibaba.cloud</groupId>
+            <artifactId>spring-cloud-starter-alibaba-sentinel</artifactId>
+        </dependency>
+
         <!-- 引入jdbc支持 -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
