+++ belita-cloud-web-api/src/main/java/com/belita/web/api/config/RESTConfiguration.java
@@ -1,4 +1,4 @@
-package com.belita.web.feign.config;
+package com.belita.web.api.config;
 
 import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.context.annotation.Bean;
