+++ belita-cloud-web-api/src/main/java/com/belita/web/api/controller/ApiOrderController.java
@@ -1,6 +1,7 @@
-package com.belita.web.feign.controller;
+package com.belita.web.api.controller;
 
 import com.beilita.common.result.JsonResult;
+import com.belita.web.api.service.OrderClient;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,6 +10,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.client.RestTemplate;
 
 import java.util.List;
+import java.util.Map;
 
 @Slf4j
 @RestController
@@ -18,8 +20,8 @@ public class ApiOrderController {
     @Autowired
     private RestTemplate restTemplate;
 
-//    @Autowired
-//    private OrderClient orderClient;
+    @Autowired
+    private OrderClient orderClient;
 
 
     @GetMapping("/findAll")
@@ -28,10 +30,10 @@ public class ApiOrderController {
         return JsonResult.success(forObject);
     }
 
-//    @GetMapping("/findInfo")
-//    public JsonResult findInfo(){
-//        Map map = orderClient.findInfo();
-//        return JsonResult.success(map);
-//    }
+    @GetMapping("/findInfo")
+    public JsonResult findInfo(Long id){
+        Map map = orderClient.findInfo(id);
+        return JsonResult.success(map);
+    }
 
 }
