+++ belita-cloud-web-api/src/main/java/com/belita/web/api/controller/IndexController.java
@@ -1,0 +1,16 @@
+package com.belita.web.api.controller;
+
+import org.springframework.cloud.context.config.annotation.RefreshScope;
+import org.springframework.web.bind.annotation.RestController;
+
+// 通过@RefreshScope 获取nacos config 配置
+@RefreshScope
+@RestController
+public class IndexController {
+
+
+
+
+
+
+}
