--- belita-cloud-web-api/src/main/java/com/belita/web/feign/OrderApiApplication.java
@@ -1,17 +1,0 @@
-package com.belita.web.feign;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-//开启feign接口扫描，指定扫描的包
-//@EnableFeignClients(basePackages = {"com.belita.web.api.service"})
-public class OrderApiApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(OrderApiApplication.class);
-    }
-
-}
