--- belita-cloud-web-feign/src/main/java/com/belita/web/feign/service/UserClient.java
@@ -1,18 +1,0 @@
-//package com.belita.web.feign.service;
-//
-//import org.springframework.cloud.openfeign.FeignClient;
-//import org.springframework.stereotype.Component;
-//import org.springframework.web.bind.annotation.GetMapping;
-//import org.springframework.web.bind.annotation.RequestParam;
-//
-//import java.util.Map;
-//
-//@Component
-//@FeignClient(name = "order-server")
-//public interface UserClient {
-//
-//    @GetMapping("/api/order/findAll")
-//    Map findInfo(@RequestParam("id") Long id);
-//
-//
-//}
