--- belita-cloud-web-feign/src/main/resources/application.yml
@@ -1,38 +1,0 @@
-##公共配置
-#server:
-#  port: 80
-#  tomcat:
-#    uri-encoding: UTF-8
-#
-#spring:
-#  application:
-#    name: service-consumer
-#  cloud: #nacos 地址配置
-#    nacos:
-#      discovery:
-#        server-addr: nacos.lc:8848
-#  #激活哪一个环境的配置文件
-#  profiles:
-#    active: dev
-#  jackson:
-#    #日期格式化
-#    date-format: yyyy-MM-dd HH:mm:ss
-#    time-zone: GMT+8
-#    serialization:
-#      #格式化输出
-#      indent_output: true
-#      #忽略无法转换的对象
-#      fail_on_empty_beans: false
-#    #设置空如何序列化
-#    defaultPropertyInclusion: NON_EMPTY
-#    deserialization:
-#      #允许对象忽略json中不存在的属性
-#      fail_on_unknown_properties: false
-#    parser:
-#      #允许出现特殊字符和转义符
-#      allow_unquoted_control_chars: true
-#      #允许出现单引号
-#      allow_single_quotes: true
-#
-#
-#
