--- belita-cloud-web-feign/src/main/resources/bootstrap.yml
@@ -1,17 +1,0 @@
-spring:
-  application:
-    name: order-feign
-  cloud:
-    nacos:
-      config:
-        enabled: true
-        server-addr: nacos.lc:8848
-        namespace: ${spring.profiles.active}
-        file-extension: yaml
-      discovery:
-        server-addr: nacos.lc:8848
-        namespace: ${spring.profiles.active}
-      username: nacos
-      password: nacos
-
-#debug: on
