+++ http/web-api.http
@@ -14,7 +14,7 @@ Accept: application/json
 
 
 ###
-GET {{baseUrl}}feign/findInfo?id=1
+GET {{baseUrl}}web/order/findInfo?id=1
 Accept: application/json
 
 ###
